Add getInstance() to NativeDgnDb
