+++ iModelCore/ECDb/ECDb/ConcurrentQueryManagerImpl.cpp
@@ -97,9 +97,9 @@ std::shared_ptr<CachedQueryAdaptor> QueryAdaptorCache::TryGet(Utf8CP ecsql, bool
 //---------------------------------------------------------------------------------------
 // @bsimethod
 //---------------------------------------------------------------------------------------
-QueryJsonAdaptor& CachedQueryAdaptor::GetJsonAdaptor() {
+ECSqlRowAdaptor& CachedQueryAdaptor::GetJsonAdaptor() {
     if (!m_adaptor) {
-         m_adaptor = std::unique_ptr<QueryJsonAdaptor>(new QueryJsonAdaptor(*m_stmt.GetECDb()));
+         m_adaptor = std::unique_ptr<ECSqlRowAdaptor>(new ECSqlRowAdaptor(*m_stmt.GetECDb()));
     }
     return *m_adaptor.get();
 }
@@ -453,7 +453,7 @@ QueryResponse::Ptr RunnableRequestBase::CreateBlobIOResponse(std::vector<uint8_t
 //---------------------------------------------------------------------------------------
 // @bsimethod
 //---------------------------------------------------------------------------------------
-QueryResponse::Ptr RunnableRequestBase::CreateECSqlResponse(std::string& resultJson, QueryProperty::List& meta, uint32_t rowCount, bool done) const {
+QueryResponse::Ptr RunnableRequestBase::CreateECSqlResponse(std::string& resultJson, ECSqlRowProperty::List& meta, uint32_t rowCount, bool done) const {
     const auto memUsed = (uint32_t)(resultJson.size());
     return std::make_shared<ECSqlResponse>(
         QueryResponse::Stats(GetCpuTime(), GetTotalTime(), memUsed,m_quota),
@@ -892,8 +892,8 @@ void QueryHelper::BindLimits(ECSqlStatement& stmt, QueryLimit const& limit) {
 //---------------------------------------------------------------------------------------
 // @bsimethod
 //---------------------------------------------------------------------------------------
-QueryProperty const& QueryProperty::List::GetPropertyInfo(std::string const& name) const {
-    static QueryProperty kNull;
+ECSqlRowProperty const& ECSqlRowProperty::List::GetPropertyInfo(std::string const& name) const {
+    static ECSqlRowProperty kNull;
     for(auto& info: *this) {
         if (info.GetJsonName() == name)
             return this->at(info.GetIndex());
@@ -914,8 +914,8 @@ std::string IJsSerializable::Stringify(StringifyFormat format) const {
 //---------------------------------------------------------------------------------------
 // @bsimethod
 //---------------------------------------------------------------------------------------
-QueryProperty::List QueryHelper::GetMetaInfo(CachedQueryAdaptor& adp, bool classIdToClassNames) {
-    QueryProperty::List props;
+ECSqlRowProperty::List QueryHelper::GetMetaInfo(CachedQueryAdaptor& adp, bool classIdToClassNames) {
+    ECSqlRowProperty::List props;
 
     return props;
 }
@@ -923,7 +923,7 @@ QueryProperty::List QueryHelper::GetMetaInfo(CachedQueryAdaptor& adp, bool class
 //---------------------------------------------------------------------------------------
 // @bsimethod
 //---------------------------------------------------------------------------------------
-void QueryProperty::ToJs(BeJsValue& val) const {
+void ECSqlRowProperty::ToJs(BeJsValue& val) const {
     val.toObject();
     val[JClass]=m_className;
     val[JAccessString]=m_accessString;
@@ -938,7 +938,7 @@ void QueryProperty::ToJs(BeJsValue& val) const {
 //---------------------------------------------------------------------------------------
 // @bsimethod
 //---------------------------------------------------------------------------------------
-void QueryProperty::List::ToJs(BeJsValue& val) const {
+void ECSqlRowProperty::List::ToJs(BeJsValue& val) const {
     val.toArray();
     for(auto it = begin(); it != end(); ++it) {
         auto newEl = val.appendValue();
@@ -948,8 +948,8 @@ void QueryProperty::List::ToJs(BeJsValue& val) const {
 //---------------------------------------------------------------------------------------
 // @bsimethod
 //---------------------------------------------------------------------------------------
-void QueryProperty::List::append(std::string className, std::string accessString, std::string jsonName, std::string name, std::string typeName, bool generated, std::string extendedType, int index) {
-    emplace_back(QueryProperty(className, accessString, jsonName, name, typeName, generated, extendedType, index));
+void ECSqlRowProperty::List::append(std::string className, std::string accessString, std::string jsonName, std::string name, std::string typeName, bool generated, std::string extendedType, int index) {
+    emplace_back(ECSqlRowProperty(className, accessString, jsonName, name, typeName, generated, extendedType, index));
 }
 
 //---------------------------------------------------------------------------------------
@@ -997,7 +997,7 @@ void QueryHelper::Execute(CachedQueryAdaptor& cachedAdaptor, RunnableRequestBase
     const auto classIdToClassNames = request.GetConvertClassIdsToClassNames();
     auto& stmt = cachedAdaptor.GetStatement();
     auto& adaptor = cachedAdaptor.GetJsonAdaptor();
-    QueryProperty::List props;
+    ECSqlRowProperty::List props;
     if (includeMetaData) {
         adaptor.GetMetaData(props ,stmt);
     }
@@ -1111,7 +1111,7 @@ void QueryHelper::ReadBlob(ECDbCR conn, RunnableRequestBase& runnableRequest) {
 // @bsimethod
 //---------------------------------------------------------------------------------------
 void QueryHelper::ExecutePing(Json::Value const& pingJson, RunnableRequestBase& runnableRequest) {
-    QueryProperty::List props;
+    ECSqlRowProperty::List props;
     props.append("", "", "id", "id", "long", false, "", 0);
 
     const auto maxMem = (int64_t)ConcurrentQueryMgr::GetConfig(runnableRequest.GetQueue().GetECDb()).GetQuota().MaxMemoryAllowed();
