+++ iModelCore/ECDb/ECDb/ECDbPch.h
@@ -64,7 +64,6 @@
 #include "ChangeIteratorImpl.h"
 #include "ChangeSummaryExtractor.h"
 #include "ConcurrentQueryManagerImpl.h"
-#include "QueryJsonAdaptor.h"
 #include "InstanceReaderImpl.h"
 #include "IntegrityChecker.h"
 #include "BuiltInVTabs.h"
